[#17] 로그인 API

+++ src/main/java/com/example/demo/src/user/model/PostLoginRes.java
@@ -8,7 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostLoginRes {
-
-    private int userIdx;
+    private long userIdx;
     private String jwt;
 }
